app/requirements.txt
flask
flask_socketio
python-dotenv
requests
openai